사용자 제거 test 실패

+++ Backend-FinancialAnaylsis/src/main/java/com/financial/analysis/service/user/impl/UserServiceImpl.java
@@ -97,6 +97,13 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUser(Long id) throws CustomUserException {
+
+        Optional<User> user = userRepository.findById(id);
+
+        if(user.isPresent()){
+            throw new CustomUserException("User not found");
+        }
+
         userRepository.deleteById(id);
     }
